# problem_5_resource/openapi.yml
openapi: 3.0.3
info:
  title: Employee & Task Management API
  description: |
    A comprehensive employee and task management system API built with Express.js and TypeScript.
    This API provides:
    - User authentication (JWT-based)
    - Employee management
    - Task management with assignment tracking
    
    All employee and task endpoints require JWT authentication.
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /employees:
    get:
      summary: List all employees
      description: Retrieve a list of all employees with pagination support
      operationId: getEmployees
      tags:
        - Employees
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of employees per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Employees retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Employee'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of employees
                        example: 50
                      page:
                        type: integer
                        description: Current page number
                        example: 1
                      totalPages:
                        type: integer
                        description: Total number of pages
                        example: 5
                      limit:
                        type: integer
                        description: Number of items per page
                        example: 10
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    post:
      summary: Create a new employee
      description: Create a new employee profile
      operationId: createEmployee
      tags:
        - Employees
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Employee created successfully"
                  data:
                    $ref: '#/components/schemas/Employee'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /employees/{id}:
    get:
      summary: Get employee by ID
      description: Retrieve a specific employee by their ID
      operationId: getEmployeeById
      tags:
        - Employees
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Employee ID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Employee retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Employee retrieved successfully"
                  data:
                    $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    put:
      summary: Update employee
      description: Update an existing employee
      operationId: updateEmployee
      tags:
        - Employees
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Employee ID
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Employee updated successfully"
                  data:
                    $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    delete:
      summary: Delete employee
      description: Delete an existing employee
      operationId: deleteEmployee
      tags:
        - Employees
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Employee ID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Employee deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  responses:
    UnauthorizedError:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Missing or invalid Authorization header"
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the employee
          example: 1
        user_id:
          type: integer
          description: Associated user account ID
          example: 1
        first_name:
          type: string
          description: Employee's first name
          example: "John"
        last_name:
          type: string
          description: Employee's last name
          example: "Doe"
        department:
          type: string
          description: Department where employee works
          example: "Engineering"
          nullable: true
        position:
          type: string
          description: Employee's job position
          example: "Software Engineer"
          nullable: true
        created_at:
          type: string
          format: date-time
          description: Timestamp when employee was created
          example: "2026-01-14T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when employee was last updated
          example: "2026-01-14T10:30:00Z"
      required:
        - id
        - user_id
        - first_name
        - last_name
    CreateEmployeeRequest:
      type: object
      properties:
        user_id:
          type: integer
          description: User account ID to link to this employee
          example: 1
        first_name:
          type: string
          description: Employee's first name
          example: "John"
          minLength: 1
        last_name:
          type: string
          description: Employee's last name
          example: "Doe"
          minLength: 1
        department:
          type: string
          description: Department where employee works
          example: "Engineering"
        position:
          type: string
          description: Employee's job position
          example: "Software Engineer"
      required:
        - user_id
        - first_name
        - last_name
    UpdateEmployeeRequest:
      type: object
      properties:
        first_name:
          type: string
          description: Employee's first name
          example: "John"
          minLength: 1
        last_name:
          type: string
          description: Employee's last name
          example: "Doe"
          minLength: 1
        department:
          type: string
          description: Department where employee works
          example: "Engineering"
        position:
          type: string
          description: Employee's job position
          example: "Senior Software Engineer"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
...